create interface Data as generic, delete comments and rename getAll to getSuppliers
