+++ services/types.ts
@@ -1,0 +1,12 @@
+export interface Data <T>{
+  data: T[],
+  total: number;
+}
+
+export type QueryParamsType = {
+  keywords: string,
+  orderBy: string,
+  orderType: string,
+  page: number,
+  limit: number,
+}
